relocating svg icons to assets directory.
