+++ src/assets/MainStageLayoutIcon/LayoutIconEmpty.tsx
