+++ src/assets/MainStageLayoutIcon/LayoutIconFullscreen.tsx
