+++ src/assets/MainStageLayoutIcon/LayoutIconStage.tsx
